app/src/test/java/com/ahmed_nezhi/listofproducts/common/NetworkUtilsTest.kt
package com.ahmed_nezhi.listofproducts.common

import android.content.Context
import android.net.ConnectivityManager
import android.net.Network
import android.net.NetworkCapabilities
import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import io.mockk.every
import io.mockk.mockk
import io.mockk.spyk
import kotlinx.coroutines.ExperimentalCoroutinesApi
import kotlinx.coroutines.test.StandardTestDispatcher
import kotlinx.coroutines.test.TestCoroutineScope
import org.junit.Assert
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.junit.runner.RunWith
import org.mockito.junit.MockitoJUnitRunner

/**
 * Create by A.Nezhi on 09/10/2023.
 */
@OptIn(ExperimentalCoroutinesApi::class)
@RunWith(MockitoJUnitRunner::class)
class NetworkUtilsTest {

    @get:Rule
    val instantTaskExecutorRule = InstantTaskExecutorRule()

    private val testDispatcher = StandardTestDispatcher()
    private val testScope = TestCoroutineScope(testDispatcher)

    private lateinit var networkUtils: NetworkUtils
    private lateinit var mockContext: Context
    private lateinit var mockConnectivityManager: ConnectivityManager
    private lateinit var mockNetwork: Network
    private lateinit var mockNetworkCapabilities: NetworkCapabilities

    @Before
    fun setup() {
        mockContext = mockk(relaxed = true)
        mockConnectivityManager = mockk(relaxed = true)
        mockNetwork = mockk(relaxed = true)
        mockNetworkCapabilities = mockk(relaxed = true)

        every { mockContext.getSystemService(Context.CONNECTIVITY_SERVICE) } returns mockConnectivityManager
        every { mockConnectivityManager.activeNetwork } returns mockNetwork
        every { mockConnectivityManager.getNetworkCapabilities(mockNetwork) } returns mockNetworkCapabilities


        networkUtils = spyk(NetworkUtils(mockContext))
    }


    @Test
    fun `test internet is available for android m and above`() {
        every { networkUtils.isAndroidMOrAbove() } returns true
        every { mockNetworkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) } returns true

        val result = networkUtils.isInternetAvailable()

        Assert.assertTrue(result)
    }

    @Test
    fun `test internet is available for android before m`() {
        every { networkUtils.isAndroidMOrAbove() } returns false
        every { mockNetworkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) } returns true

        val result = networkUtils.isInternetAvailable()

        Assert.assertFalse(result)
    }

    @Test
    fun `test internet is not available for android m and above`() {
        every { networkUtils.isAndroidMOrAbove() } returns true
        every { mockNetworkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) } returns false

        val result = networkUtils.isInternetAvailable()

        Assert.assertFalse(result)
    }
    @Test
    fun `test internet is not available for android before m`() {
        every { networkUtils.isAndroidMOrAbove() } returns false
        every { mockNetworkCapabilities.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET) } returns false

        val result = networkUtils.isInternetAvailable()

        Assert.assertFalse(result)
    }


}
